+++ taruni/css/todo.css
@@ -1,72 +1,82 @@
-body,button{font-family: Rockwell;}
-button{
+body,button {font-family: Rockwell;}
+button {
   background-color: #00cc99;
   color:#ffffff;
   margin-left:5px;
   height:30px;
 }
-section{
-  margin:35px;
+section {
+  margin-top:35px;
+  margin-left:350px;
+  margin-right:350px;
   background-color: #e6e6e6;
   padding-bottom:20px;
+  height: 500px;
 }
-h1{
+h1 {
   color:#00cc99;
   display:inline;
 }
-ul{
+ul {
   list-style: none;
   padding:0;
   margin-top:0;
   margin-bottom:0;
   margin-right:0;
-  width:500px;
+  /*width:500px;*/
 }
-li{
+li {
   border:1px solid #ccc;
   background: #eee;
   padding:5px 10px;
   color:#527a7a;
   background-color: #ffffcc;
-  width:150%;
+  /*width:150%;*/
   margin-left:10px;
+  margin-right:10px;
 }
-li span{
+li span {
   padding-left: 10px;
   cursor: default;
 }
-.checked{
+.checked {
   text-decoration: line-through;
   font-weight: bold;
   color:#40bf80;
 }
-@media (max-width: 609px) {
+@media (max-width: 382px) {
   section{
-         overflow: hidden;
-         min-width:269px;
+   overflow: hidden;
+   min-width:345px;
+  }
 }
-}
-@media (max-width: 605px) {
+@media (max-width: 382px) {
   li{
          overflow: hidden;
-         min-width:269px;
+         min-width:345px;
+  }
 }
-}
-#taskbox{
+#taskbox {
   width:250px;
   height:25px;
 }
-input{
+input {
   color:#527a7a;
   background-color: #ffffcc;
   margin-left:10px;
   margin-bottom:10px;
+  border:1px solid lightgray;
 }
-input:focus{
+input:focus {
   background-color: #ffffff;
 }
-img{
+img {
   margin-top:5px;
   width:35px;
   margin-left:10px;
 }
+.strike-through {
+  text-decoration: line-through;
+}
+.undone {background-color: #00cc99;}
+.done {background-color:  #ffdb4d;}
