+++ taruni/html/todo.html
@@ -4,6 +4,7 @@
 	<title>To Do</title>
 	<link rel="icon" href="..\images\notepad.png">
 	<link rel="stylesheet" type="text/css" href="..\css\todo.css">
+	<link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
 </head>
 <body>
 	<section id="todolistitems">
